Handle files that contain a exif orientation metadata tag
To make sure the correct region gets selected, we add a processing step that applies the metadata tag to the actual image data, then removes the metadata tag from the file object.

Otherwise the plug would work on the "unflipped" image data when selection the requested region, which in the example would be starting at the top right corner, selecting down and right.

Fixes issue #6.